Updated the GameCamera.
Added a RotatingPlatform that moves about an origin point.

Added a global_assets folder to be loaded in every level. (Has player sprites)

Split grass textures into 2 parts that allow the grass to be infront and behind the player relative to the position of the player.

Added some more properties to the Animation class that allows animations to be slowed down or sped up.

Added a BasicMovingPlatform that allows objects to pass through the bottom and will latch any objects ontop of it on to it.

Fixed the StaticCollider so it has an onCollision method that can cancel a collision and takes in the properties of the collision.

Added a GUIButton that can perform actions when clicked.

Updated the Renderer to work a little bit smoother.

Added a GetElapsedTime() method to GameTime.